rpi-gpio/requirements.txt
RPi.bme280==0.2.3
smbus2==0.3.0
flask
